fix. 주문 생성 API 추가

+++ order/src/main/kotlin/lian/sample/domain/entity/DiscountEntity.kt
@@ -1,6 +1,7 @@
 package lian.sample.domain.entity
 
 import jakarta.persistence.*
+import java.math.BigDecimal
 
 @Entity
 @Table(name = "DISCOUNT", schema = "order")
@@ -14,5 +15,5 @@ data class DiscountEntity(
   val discountType: String,
 
   @Column(name = "DISCOUNT_VALUE", nullable = false)
-  val discountValue: Double
+  val discountValue: BigDecimal
 )
